--- a/src/terrain_3d_data.h
+++ b/src/terrain_3d_data.h
@@ -104,6 +104,7 @@ public:
 	bool has_region(const Vector2i &p_region_loc) const { return get_region_id(p_region_loc) != -1; }
 	bool has_regionp(const Vector3 &p_global_position) const { return get_region_idp(p_global_position) != -1; }
 	Ref<Terrain3DRegion> get_region(const Vector2i &p_region_loc) const;
+	Terrain3DRegion *get_region_ptr(const Vector2i &p_region_loc) const;
 	Ref<Terrain3DRegion> get_regionp(const Vector3 &p_global_position) const;
 
 	void set_region_modified(const Vector2i &p_region_loc, const bool p_modified = true);
@@ -234,6 +235,13 @@ inline int Terrain3DData::get_region_idp(const Vector3 &p_global_position) const
 
 inline Ref<Terrain3DRegion> Terrain3DData::get_region(const Vector2i &p_region_loc) const {
 	return _regions.get(p_region_loc, Ref<Terrain3DRegion>());
+}
+
+inline Terrain3DRegion *Terrain3DData::get_region_ptr(const Vector2i &p_region_loc) const {
+	if (has_region(p_region_loc)) {
+		return cast_to<Terrain3DRegion>(_regions[p_region_loc]);
+	}
+	return nullptr;
 }
 
 inline Ref<Terrain3DRegion> Terrain3DData::get_regionp(const Vector3 &p_global_position) const {