--- a/src/terrain_3d_editor.h
+++ b/src/terrain_3d_editor.h
@@ -86,7 +86,7 @@ private:
 	uint64_t _last_pen_tick = 0;
 
 	void _send_region_aabb(const Vector2i &p_region_loc, const Vector2 &p_height_range = Vector2());
-	Ref<Terrain3DRegion> _operate_region(const Vector2i &p_region_loc);
+	Terrain3DRegion *_operate_region(const Vector2i &p_region_loc);
 	void _operate_map(const Vector3 &p_global_position, const real_t p_camera_direction);
 	MapType _get_map_type() const;
 	bool _is_in_bounds(const Point2i &p_pixel, const Point2i &p_size) const;