Admin edit
* Optimizes Terrain3DData, eg. set/get_pixel(), get_height() and more by replacing creation of Refs with pointer usage
* Meshing will be optimized in or after #622 
* Checks more pointers before using
* Consolidates nullptr checks using implicit boolean conversion
* Fixes some improper usage 


Results
```
10M get_pixel(): 2836ms => 1713ms, +39.6%
1M get_height() interpolated: 1571ms => 982ms, +37.5%
bake_mesh 3x1024: 40765ms => 34750ms, +14.8%
```

-------
This PR should cut down navigation mesh baking time by about half. Terrain3D::_generate_triangle_pair uses fewer control map lookups (-6 when p_require_nav is false, -2 if true), and Terrain3DStorage::get_pixel uses a modified private version of get_map_region that returns an Image* and skips converting it to a godot::Ref, reducing the time spent on get_pixel by about 1/3rd. This is based on the assumption that nobody will put anything other than godot::Image in the height/control/color map arrays, which should be true because they're TypedArrays and they're not directly accessible from outside the extension.

The modified version of get_map_region introduces some code duplication, and it might be good to use this modified version in other places where a Ref really isn't needed, so this PR isn't as clean as it perhaps could've been. I'm personally unlikely to even use it in my project, as it should be much faster to just directly sample the physics shapes I'm making, which can also be generated in a thread, unlike Terrain3D.