@TokisanGames 2024-08-19
get_map_region has been dropped. You can see here in set/get_pixel how the general API focus is changing towards selecting regions by location (fixed) rather than ids (mutable) and working with Regions. So the ptr function probably needs to be moved to Terrain3DRegion.

https://github.com/TokisanGames/Terrain3D/pull/374/commits/6d77480e1cf382c89a5284c6cf160a9c34d9b560#diff-f5a6479fa5fc2f89383edd16b70ed900e9b288af17f42366f3ac1bdb5a19306a

Although we are keeping the typedarrays referenced by ID, which are updated by update_maps, and then sent to the shader. So if more optimal we could undo these changes to set/get_pixel and reference the maps directly bypassing regions and refs.

Read the comments here
https://github.com/TokisanGames/Terrain3D/blob/6d77480e1cf382c89a5284c6cf160a9c34d9b560/src/terrain_3d_storage.h#L48-L79

I'd recommend just waiting until I finish my PR before spending any more time on it, but wanted to share the information to get you thinking about it.