@TokisanGames 2025-03-02
I'm taking this over and expanding the optimizations throughout. One already helped improve #599.