@Xtarsia 2025-03-06
Everything looks good, painting noticably faster!

I tested painting texture with 1000m brush size, before this: 4.3s, After 3.1s 🎉